app docker ignore e readme

+++ app/models/sale.py
@@ -1,4 +1,6 @@
-from sqlalchemy import Date, Float, Integer, String
+from decimal import Decimal
+
+from sqlalchemy import Date, Integer, Numeric, String
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.db.sqlite import Base
@@ -11,6 +13,6 @@ class Sale(Base):
     product_name: Mapped[str] = mapped_column(String(255), nullable=False)
     category: Mapped[str] = mapped_column(String(100), nullable=False, index=True)
     quantity: Mapped[int] = mapped_column(Integer, nullable=False)
-    unit_price: Mapped[float] = mapped_column(Float, nullable=False)
+    unit_price: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=False)
     sale_date: Mapped[Date] = mapped_column(Date, nullable=False, index=True)
